fix: apply a patch to fix "field formatter class" in layout builder
This involves the following steps:
1. We are able to remove the first original patch by advancing the SHA
we are using, because the first patch was merged into the module.
2. We apply a new patch to fix an issue with field formatter class and
layout builder. This fix could be merged into the module or made
unnecessary by future core changes, but in the mean time it appears to
be a stable workaround.

See https://www.drupal.org/project/field_formatter_class/issues/3046372
for more details.