feat: use padding to adjust width for caption add auto height
Using the padding for the width of the caption allows us to keep the caption container full width for the gradient while controlling the length of the caption text.

Adding height: auto addresses hb-raised-cards bug with captions and gives it possible height decleration with any possible future utility classes that alter the cards image height.