STN-628:  Caption/credit text styles limit (#790)
* feat:(STN-628) caption/credit text style limit

I removed the widths, the clear and the margin on the individual elements so that they can display inline with each other rather than separate cleared lines of elements. I added a small margin to the right so elements appear to have spacing when they normally wouldn't. I additionally added padding to the container to account for the margin removal from the child elements. While also adding flex and flex-end to the container so the float right elements to not appear in reverse order on the caption.

* chore: compile css