fix: adding overflow-x: hidden to html selector as overflow-x: hidden on the body tag ( mobile browsers ) isn't necessarily respected.
The color band component is able to be displayed full-width and it is
currently relying on overflow-x: hidden on the body to prevent overflow
issues, but after testing on mobile devices you can see that this simply
doesn't work.

Add overflow-x: hidden; to html selector and this seems to resolve the
issue.